% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_mareg}
\alias{list_mareg}
\title{List files/directories in the mareg root directory}
\usage{
list_mareg(path = mareg_path(), ...)
}
\arguments{
\item{path}{char, the path to list}

\item{...}{other arguments for \code{\link[base]{list.files}}}
}
\value{
character vector
}
\description{
This is a wrapper around \code{\link[base]{list.files}}
}
