% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mareg-package.R
\docType{package}
\name{mareg}
\alias{mareg}
\alias{mareg-package}
\title{Fetch and manage haedat datasets.}
\description{
A small alternative to mregions package for accessing MarineRegions.org data.
}
\seealso{
\href{Marine Regions}{https://www.marineregions.org/downloads.php}
}
\keyword{internal}
