% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{mareg_root}
\alias{mareg_root}
\title{Create or retrieve the root data directory}
\usage{
mareg_root(new_dir = rappdirs::user_data_dir("marineregions"))
}
\arguments{
\item{new_dir}{char, used to create a new directory the first time this is
run}
}
\value{
the root data path
}
\description{
First search for ~/.marineregions which may contain the desired path
If absent search for rappdirs::user_data_dir("marineregions")
If absent search for rappdirs::site_data_dir("marineregions")
If absent create the path specified by new_dir argument
}
\details{
If you want to specify a non-rappdirs standard path, then create a
~/mareg file with one line specifying the desired data path. You must
create this path yourself.
}
