% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_eez}
\alias{read_eez}
\title{Read EEZ geography}
\usage{
read_eez(version = "v11", res = c("high", "low")[2], boundaries = FALSE)
}
\arguments{
\item{version}{char, by default "v11"}

\item{res}{char, one of "low" (default) or "high"}

\item{boundaries}{logical, by default FALSE to read all of the data}
}
\value{
sf table
}
\description{
Read EEZ geography
}
